CODE_06062025/retrieval/kite/requirements.txt
mlflow
cloudpickle
databricks-connect
databricks-vectorsearch
google-cloud-storage
ipykernel
langchain-community
langchain
numpy
pandas
pyarrow
pydantic
pyspark
databricks-langchain